add image_encoding_network
